+++ .eslintrc.js
@@ -1,0 +1,24 @@
+module.exports = {
+  env: {
+    es2021: true,
+    node: true,
+  },
+  extends: ['airbnb-typescript/base', 'plugin:@typescript-eslint/recommended'],
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 2018,
+    sourceType: 'module',
+    project: './tsconfig.json',
+  },
+  rules: {
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
+    'no-param-reassign': 'off',
+    'no-underscore-dangle': 'off',
+    'consistent-return': 'off',
+    'no-console': 'off',
+    'import/prefer-default-export': 'off',
+    '@typescript-eslint/no-non-null-assertion': 'off',
+    'class-methods-use-this': 'off',
+    '@typescript-eslint/naming-convention': 'off'
+  },
+};