+++ jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  transform: {
+    '^.+\\.ts?$': 'ts-jest',
+  },
+  setupFiles: ['dotenv/config'],
+  testEnvironment: 'node',
+  coveragePathIgnorePatterns: ['node_modules'],
+  coverageReporters: ['text', 'lcov', 'clover', 'html'],
+};