+++ src/config/config.ts
@@ -1,0 +1,8 @@
+/*
+
+Copyright (c) 2019 - present AppSeed.us
+
+*/
+export default {
+  webURL: 'https://localhost:5000/',
+};