+++ src/models/user.ts
@@ -1,0 +1,19 @@
+import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity()
+export default class User {
+  @PrimaryGeneratedColumn('uuid')
+  id!: string;
+
+  @Column({ type: 'text', nullable: false })
+  username!: string;
+
+  @Column({ type: 'text', nullable: false })
+  email!: string;
+
+  @Column({ type: 'text', nullable: false })
+  password!: string;
+
+  @Column({ type: 'datetime', default: () => 'CURRENT_TIMESTAMP' })
+  date?: string;
+}